div first then multiply

+++ executor/aggfuncs/func_percentile.go
@@ -39,7 +39,7 @@ var (
 
 func percentile(data sort.Interface, percent int) int {
 	// Ordinal rank k = Ceil(P / 100 * N)
-	k := int(math.Ceil(float64(data.Len()*percent) / 100))
+	k := int(math.Ceil(float64(data.Len()) * (float64(percent) / 100)))
 	return selection.Select(data, k)
 }
